+++ apps/decodex/src/mcp/http/handler/request/cors.rs
@@ -1,0 +1,23 @@
+use super::McpHttpHandler;
+use crate::mcp::http::{message::McpHttpRequest, security};
+
+impl McpHttpHandler {
+	pub(super) fn allowed_cors_origin(
+		&self,
+		request: &McpHttpRequest,
+	) -> std::result::Result<Option<String>, ()> {
+		let Some(origin) = request.header("Origin") else {
+			return Ok(None);
+		};
+
+		if security::mcp_http_origin_is_allowed(
+			origin,
+			self.listen_address.as_deref(),
+			self.allowed_origins.as_slice(),
+		) {
+			Ok(Some(origin.to_owned()))
+		} else {
+			Err(())
+		}
+	}
+}
